Moved non-console-app logic into the Core project
Moves core AI logic and domain models into the Core project to better reflect clean architecture principles.

Renames and relocates configuration and logging extensions to a more appropriate Infrastructure folder.

Updates code style guidelines in the documentation to prefer target-typed new expressions over var for improved readability, except within foreach loops.

Switches from Serilog to ILogger/ILoggerFactory for logging abstraction, enhancing flexibility and testability.
