+++ AiTableTopGameMaster.ConsoleApp/Infrastructure/LoggingExtensions.cs
@@ -3,7 +3,7 @@ using Microsoft.Extensions.Logging;
 using Serilog;
 using Serilog.Filters;
 
-namespace AiTableTopGameMaster.ConsoleApp.Helpers;
+namespace AiTableTopGameMaster.ConsoleApp.Infrastructure;
 
 public static class LoggingExtensions
 {
