--- AiTableTopGameMaster.ConsoleApp/Models/ModelInfo.cs
@@ -1,3 +1,0 @@
-namespace AiTableTopGameMaster.ConsoleApp.Cores;
-
-public abstract record ModelInfo;
