+++ AiTableTopGameMaster.Core/Domain/Adventure.cs
