+++ AiTableTopGameMaster.Core/Domain/Character.cs
