+++ AiTableTopGameMaster.Core/Domain/Encounter.cs
