+++ AiTableTopGameMaster.Core/Domain/Location.cs
