+++ AiTableTopGameMaster.Core/Models/ModelInfo.cs
@@ -1,0 +1,3 @@
+namespace AiTableTopGameMaster.Core.Models;
+
+public abstract record ModelInfo;
