+++ AiTableTopGameMaster.Core/Models/OllamaModelInfo.cs
@@ -1,3 +1,3 @@
-namespace AiTableTopGameMaster.ConsoleApp.Cores;
+namespace AiTableTopGameMaster.Core.Models;
 
 public record OllamaModelInfo(string Name, Uri Url) : ModelInfo;
