+++ AiTableTopGameMaster.Core/Settings/OllamaSettings.cs
