+++ AiTableTopGameMaster.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 17.13.35931.197 d17.13
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AiTableTopGameMaster.ConsoleApp", "AiTableTopGameMaster.ConsoleApp\AiTableTopGameMaster.ConsoleApp.csproj", "{A20BC0AF-F198-4EFA-B9D2-32F593430A63}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AiTableTopGameMaster.Domain", "AiTableTopGameMaster.Domain\AiTableTopGameMaster.Domain.csproj", "{529004CE-FBD2-411D-AB90-412B1A1AF9D6}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AiTableTopGameMaster.Core", "AiTableTopGameMaster.Core\AiTableTopGameMaster.Core.csproj", "{50AD4885-B9A1-4926-8603-D503D339A68C}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AiTableTopGameMaster.Tests", "AiTableTopGameMaster.Tests\AiTableTopGameMaster.Tests.csproj", "{1B9F8C8D-2E3A-4F5C-8A1B-9D5E6F7A8B2C}"
@@ -27,10 +25,6 @@ Global
 		{A20BC0AF-F198-4EFA-B9D2-32F593430A63}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A20BC0AF-F198-4EFA-B9D2-32F593430A63}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A20BC0AF-F198-4EFA-B9D2-32F593430A63}.Release|Any CPU.Build.0 = Release|Any CPU
-		{529004CE-FBD2-411D-AB90-412B1A1AF9D6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{529004CE-FBD2-411D-AB90-412B1A1AF9D6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{529004CE-FBD2-411D-AB90-412B1A1AF9D6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{529004CE-FBD2-411D-AB90-412B1A1AF9D6}.Release|Any CPU.Build.0 = Release|Any CPU
 		{50AD4885-B9A1-4926-8603-D503D339A68C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{50AD4885-B9A1-4926-8603-D503D339A68C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{50AD4885-B9A1-4926-8603-D503D339A68C}.Release|Any CPU.ActiveCfg = Release|Any CPU
